@@ -1,5 +1,5 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
-import {FormBuilder, Validators} from "@angular/forms";
+import {UntypedFormBuilder, Validators} from "@angular/forms";
 import {MessageService} from "../../services/messages.service";
 import {MatDialog} from "@angular/material/dialog";
 import {UsersService} from "../../services/users.service";
@@ -32,7 +32,7 @@ export class UsersComponent implements OnInit {
         private usersService: UsersService,
         private dependenciesService: DependenciesService,
         private messagesService: MessageService,
-        private formBuilder: FormBuilder,
+        private formBuilder: UntypedFormBuilder,
         public matDialog: MatDialog,
     ) {
     }
